Cleanup - Remove circular dependency of mapping by moving all to `mappings.py` - Constants moved to `constants.py`
