+++ pyard/misc.py
@@ -1,17 +1,9 @@
 # List of expression characters
 import pathlib
-import re
 import tempfile
 from typing import List
 
-HLA_regex = re.compile("^HLA-")
-
-VALID_REDUCTION_TYPES = ["G", "P", "lg", "lgx", "W", "exon", "U2"]
-expression_chars = ["N", "Q", "L", "S"]
-# List of P and G characters
-PandG_chars = ["P", "G"]
-
-DEFAULT_CACHE_SIZE = 1_000
+from pyard.constants import VALID_REDUCTION_TYPES, expression_chars, PandG_chars
 
 
 def get_n_field_allele(allele: str, n: int, preserve_expression=False) -> str:
