+++ scripts/pyard-status
@@ -26,6 +26,7 @@ import os
 import re
 
 import pyard
+import pyard.mappings
 from pyard import db, data_repository
 from pyard.misc import get_data_dir
 
@@ -93,8 +94,8 @@ if __name__ == "__main__":
             print(f"|{'Table Name':20}|{'Rows':20}|")
             print(f"|{'-' * 41}|")
             for table in (
-                data_repository.ars_mapping_tables
-                + data_repository.code_mapping_tables
+                pyard.mappings.ars_mapping_tables
+                + pyard.mappings.code_mapping_tables
                 + ["mac_codes"]
             ):
                 if db.table_exists(db_connection, table):
